requirements.txt
asgiref==3.8.1
decorator==5.1.1
Django==5.1.4
django-cors-headers==4.6.0
django-mysql==4.15.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ffmpeg-python==0.2.0
future==1.0.0
imageio==2.36.1
imageio-ffmpeg==0.5.1
mysqlclient==2.2.7
numpy==2.2.1
pillow==10.4.0
proglog==0.1.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
setuptools==75.7.0
sqlparse==0.5.3
tqdm==4.67.1